# openapi/spec.yaml
openapi: '3.0.3'
info:
  title: Code-Job API
  description: >
    Bilingual (Arabic/English) AI-native commercial-contracts workspace API.
    Default locale: Arabic (RTL). All endpoints require authentication
    unless explicitly marked otherwise.
  version: 0.1.0
  contact:
    name: Code-Job Engineering
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://staging.example.com/api/v1
    description: Staging (Cloudways — URL TBD after D-01)
paths:
...
  /api/v1/matters/{matter_id}/counterparties:
    post:
      summary: Attach counterparty to matter
      operationId: attachCounterparty
      tags: [Matters]
      parameters:
        - name: matter_id
          in: path
          required: true
          schema: { type: string, format: ulid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                contact_id: { type: string, format: ulid }
                representing: { type: string, enum: [we_represent, they_represent, no_counsel] }
              required: [contact_id]
      responses:
        '201':
          description: Counterparty attached
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v1/matters/{matter_id}/counterparties/{contact_id}:
    delete:
      summary: Detach counterparty from matter
      operationId: detachCounterparty
      tags: [Matters]
      parameters:
        - name: matter_id
          in: path
          required: true
          schema: { type: string, format: ulid }
        - name: contact_id
          in: path
          required: true
          schema: { type: string, format: ulid }
      responses:
        '204':
          description: Counterparty detached
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/matters/{matter_id}/lawyers:
    post:
      summary: Attach lawyer to matter
      operationId: attachLawyer
      tags: [Matters]
      parameters:
        - name: matter_id
          in: path
          required: true
          schema: { type: string, format: ulid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id: { type: string, format: ulid }
                role: { type: string, maxLength: 50 }
              required: [user_id]
      responses:
        '201':
          description: Lawyer attached
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/matters/{matter_id}/lawyers/{user_id}:
    delete:
      summary: Detach lawyer from matter
      operationId: detachLawyer
      tags: [Matters]
      parameters:
        - name: matter_id
          in: path
          required: true
          schema: { type: string, format: ulid }
        - name: user_id
          in: path
          required: true
          schema: { type: string, format: ulid }
      responses:
        '204':
          description: Lawyer detached
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
              errors:
                type: object
                additionalProperties:
                  type: array
                  items:
                    type: string
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
security:
  - bearerAuth: []